chore: separate goods package
