+++ pkg/addons/applier.go
@@ -21,14 +21,12 @@ import (
 	"github.com/replicatedhq/embedded-cluster/pkg/addons/seaweedfs"
 	"github.com/replicatedhq/embedded-cluster/pkg/addons/velero"
 	"github.com/replicatedhq/embedded-cluster/pkg/helpers"
-	"github.com/replicatedhq/embedded-cluster/pkg/kotscli"
 	"github.com/replicatedhq/embedded-cluster/pkg/kubeutils"
 	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
 	"github.com/replicatedhq/embedded-cluster/pkg/spinner"
+	"github.com/replicatedhq/embedded-cluster/pkg/support"
 	"github.com/replicatedhq/troubleshoot/pkg/apis/troubleshoot/v1beta2"
 )
-
-const SpecDataKey = "support-bundle-spec"
 
 // AddOn is the interface that all addons must implement.
 type AddOn interface {
@@ -89,7 +87,7 @@ func (a *Applier) Outro(ctx context.Context, k0sCfg *k0sv1beta1.ClusterConfig, e
 		return err
 	}
 
-	err = kotscli.CreateHostSupportBundle()
+	err = support.CreateHostSupportBundle()
 	if err != nil {
 		logrus.Warnf("failed to create host support bundle: %v", err)
 	}
