+++ pkg/goods/goods.go
@@ -8,7 +8,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"strings"
 )
 
 var (
@@ -36,23 +35,4 @@ func K0sBinarySHA256() (string, error) {
 		return "", fmt.Errorf("unable to copy embedded k0s binary: %w", err)
 	}
 	return hex.EncodeToString(hasher.Sum(nil)), nil
-}
-
-func GetSupportBundleSpec(name string) ([]byte, error) {
-	entries, err := supportfs.ReadDir("support")
-	if err != nil {
-		return nil, fmt.Errorf("unable to read embedded-cluster support dir: %w", err)
-	}
-	for _, entry := range entries {
-		srcpath := fmt.Sprintf("support/%s", entry.Name())
-		if !strings.Contains(entry.Name(), name) {
-			continue
-		}
-		srcfile, err := supportfs.ReadFile(srcpath)
-		if err != nil {
-			return nil, fmt.Errorf("unable to read asset: %w", err)
-		}
-		return srcfile, nil
-	}
-	return nil, fmt.Errorf("unable to find support file %s", name)
 }
