+++ pkg/support/static/host-support-bundle-remote.yaml
