+++ pkg/upgrade/upgrade.go
@@ -17,8 +17,8 @@ import (
 	"github.com/replicatedhq/embedded-cluster/operator/pkg/release"
 	"github.com/replicatedhq/embedded-cluster/pkg/config"
 	"github.com/replicatedhq/embedded-cluster/pkg/helpers"
-	"github.com/replicatedhq/embedded-cluster/pkg/kotscli"
 	"github.com/replicatedhq/embedded-cluster/pkg/kubeutils"
+	"github.com/replicatedhq/embedded-cluster/pkg/support"
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -71,7 +71,7 @@ func Upgrade(ctx context.Context, in *clusterv1beta1.Installation) error {
 		return fmt.Errorf("registry migration status: %w", err)
 	}
 
-	err = kotscli.CreateHostSupportBundle()
+	err = support.CreateHostSupportBundle()
 	if err != nil {
 		logrus.Warnf("Failed to upgrade host support bundle: %v", err)
 	}
